app/requirements.txt
fastapi
uvicorn
spacy==3.7.5
spacy-transformers
python-dotenv
sentry-sdk[fastapi]